Taha AIML Task/backend/requirements.txt
Flask 
psycopg2-binary 
google-cloud-speech 
ffmpeg-python 
streamlit
flask_sqlalchemy
numpy==1.26.4